requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
bs4==0.0.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
Deprecated==1.2.18
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
feedparser==6.0.11
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
narwhals==1.42.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.2.2
packaging==24.2
pandas==2.3.0
patsy==1.0.1
pillow==11.2.1
plotly==6.1.2
praw==7.8.1
prawcore==2.4.0
preshed==3.0.10
protobuf==6.31.1
psycopg2==2.9.10
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pymisp==2.5.12
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
statsmodels==0.14.4
stopwords==1.0.1
streamlit==1.45.1
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
tweepy==4.15.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.4.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
websocket-client==1.8.0
wrapt==1.17.2
